+++ craftassist/test/base_craftassist_test_case.py
@@ -9,7 +9,7 @@ from typing import List, Sequence, Dict
 from droidlet.lowlevel.minecraft.build_utils import to_relative_pos
 from droidlet.memory.craftassist.mc_memory_nodes import VoxelObjectNode
 from droidlet.lowlevel.minecraft.mc_util import XYZ, Block, IDM
-from droidlet.perception import yaw_pitch
+from droidlet.perception.craftassist.rotation import yaw_pitch
 from droidlet.dialog.dialogue_objects import AwaitResponse
 
 from .fake_agent import FakeAgent, FakePlayer
