+++ droidlet/perception/craftassist/voxel_models/subcomponent_classifier.py
@@ -7,8 +7,8 @@ from multiprocessing import Queue, Process
 import sys
 import os
 from droidlet.memory.craftassist.mc_memory_nodes import InstSegNode
-from droidlet.perception import all_nearby_objects
-from droidlet.perception import get_bounds
+from droidlet.perception.craftassist.heuristic_perception import all_nearby_objects
+from droidlet.perception.craftassist.shapes import get_bounds
 
 from droidlet.lowlevel.minecraft import build_utils as bu
 from .semantic_segmentation.semseg_models import SemSegWrapper
